remove unused transport stuff
